Fix audit findings: SSRF guard, confined writes, CI, and doc honesty
Addresses the sixdesk audit directive (baseline 6bab30f).

Blockers:
- README no longer documents `--translate-to`; it does not exist and the
  example was replaced with working commands plus an explicit note that
  translation is unimplemented.
- `ruff check .` now exits 0 (was 19 errors). Fixed unused imports/vars,
  narrowed a blind catch, removed a redundant int() cast, made the two
  subprocess.run call sites' check=False explicit, and documented the one
  intentional last-resort catch in runner.py.

Majors:
- CI added (.github/workflows/ci.yml): ruff + pytest across Python
  3.10-3.13 on push/PR. Previously zero workflows, so every test was an
  orphan and silent regression was undetectable.
- SSRF guard in sources/detect.py. Loopback, link-local (incl.
  169.254.169.254), private, CGNAT, multicast and reserved ranges are
  rejected by literal IP, blocked hostname, and by resolving the host and
  checking every address. Bracketed IPv6 literals are parsed correctly
  (a naive netloc split silently bypassed the check). The RFC 2544
  benchmarking range is deliberately NOT blocked: Python calls it private,
  but local DNS filters resolve real hostnames there.
- Output paths confined (core/paths.py). Caller-supplied output_dir is
  resolved against TEXTFLOWKIT_OUTPUT_ROOT (default: cwd). ".." escapes,
  absolute paths outside the root, and symlinks that escape are rejected
  with an actionable error (422 on HTTP). This is a separate defect from
  the SSRF guard and was fixed separately.
- Dead `--speaker-labels` flag removed end to end. It was accepted,
  threaded through four layers, and discarded. Diarization is a separate
  product decision, so the flag was removed rather than stubbed. The
  Segment.speaker field and renderer support are retained as the data
  contract.
- Harness claim corrected. README and docs/adapters.md now state that
  transport support was established by reading configs (static), not by
  live MCP calls from each harness.

Minors:
- Install docs now say the package is not on PyPI and give a verified
  source-install path.
- CONTRIBUTING.md added with the exact CI commands.
- SECURITY.md added: reporting path, supported versions, and an explicit
  threat model covering what is and is not defended.
- .pre-commit-config.yaml added.

Tests: 74 passing (was 59). New suites for the SSRF guard (13 blocked
vectors, IPv6 parsing, benchmarking-range exemption) and path confinement
(10 cases). Both new guards are mutation-verified: disabling each turns
the corresponding tests red.

Not done, deliberately: no push, tag, or release.

+++ src/textflowkit/core/timeutil.py
@@ -7,7 +7,7 @@ def srt_timestamp(seconds: float) -> str:
     """HH:MM:SS,mmm (SRT uses a comma)."""
     if seconds < 0:
         seconds = 0.0
-    ms_total = int(round(seconds * 1000))
+    ms_total = round(seconds * 1000)
     h, rem = divmod(ms_total, 3_600_000)
     m, rem = divmod(rem, 60_000)
     s, ms = divmod(rem, 1000)
@@ -17,3 +17,4 @@ def srt_timestamp(seconds: float) -> str:
 def vtt_timestamp(seconds: float) -> str:
     """HH:MM:SS.mmm (WebVTT uses a period)."""
     return srt_timestamp(seconds).replace(",", ".")
+
